requirements.txt
# Hardware dependencies
pyudev>=0.24
Pillow>=10.0

# Neopixel LED support
adafruit-circuitpython-neopixel>=6.3
adafruit-circuitpython-led-animation>=2.7
rpi_ws281x>=5.0  # For Raspberry Pi

flask>=2.3
flask-socketio>=5.3
python-socketio>=5.9

# Optional: SVG rendering
# cairosvg>=2.7